requirements.txt
pypdf
pdf2image
pdfplumber
pdf-annotate
pypdf==4.0.1
fpdf
ipython
PyPDF2==3.0.1
pypdfium2==3.21.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
matplotlib
streamlit==1.30.0
scikit-learn
pandas
numpy==1.23.0
tqdm
scikit-learn
tensorboard
tensorboard-data-server
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
transformers
Pillow==9.3.0
-e git+https://github.com/mindee/doctr.git@31f05c8320aea3db757c9be2d1589c8ee11fe2f9#egg=python_doctr
